Move homepage styling for next
